requirements.txt
fastapi
uvicorn
python-dotenv
msal
httpx
python-jose[cryptography]
python-multipart
pydantic
pydantic-settings
starlette
redis
slowapi
sqlalchemy
psycopg2-binary
pymysql
psycopg2-binary
asyncpg   # pip install psycopg2-binary asyncpg
openai
google-generativeai
azure-storage-blob
aiofiles
python-magic
pillow
tiktoken